requirements.txt
protobuf==5.29.3
PyYAML==6.0.3
setuptools==80.9.0
six==1.17.0